+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Deprecate default instantiation of TestUtils in TestCase - `$this->utils` will be removed in next major release. Do your own instantiation when needed.
 - Deprecate `TestUtils::setSelect2Value()` method. Use directly the new `Select2` component.
 - Deprecate `TestUtils::getFixturePath()` method. Use `Facebook\WebDriver\Remote\FileDetector` instead.
+- Deprecate `TestUtils::sleep()` method. Use `AbstractTestCase::sleep()` method instead.
+
 
 ## 2.0.1 - 2016-12-01
 ### Fixed